* Add a bunch of Mists enchants. * Add Alchemist's Flask. * Fix some warnings.
git-svn-id: https://simulationcraft.googlecode.com/svn/branches/mop@12617 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_dot.cpp
@@ -35,7 +35,7 @@ void dot_t::extend_duration( int extra_ticks, bool cap, uint32_t state_flags )
   if ( ! ticking )
     return;
 
-  if ( state_flags == -1 ) state_flags = action -> snapshot_flags;
+  if ( state_flags == ( uint32_t ) -1 ) state_flags = action -> snapshot_flags;
 
   // Make sure this DoT is still ticking......
   assert( tick_event );
@@ -75,7 +75,7 @@ void dot_t::extend_duration_seconds( timespan_t extra_seconds, uint32_t state_fl
   if ( ! ticking )
     return;
 
-  if ( state_flags == -1 ) state_flags = action -> snapshot_flags;
+  if ( state_flags == ( uint32_t ) -1 ) state_flags = action -> snapshot_flags;
 
   // Make sure this DoT is still ticking......
   assert( tick_event );
@@ -162,7 +162,7 @@ void dot_t::refresh_duration( uint32_t state_flags )
   if ( ! ticking )
     return;
 
-  if ( state_flags == -1 ) state_flags = action -> snapshot_flags;
+  if ( state_flags == ( uint32_t ) -1 ) state_flags = action -> snapshot_flags;
 
   // Make sure this DoT is still ticking......
   assert( tick_event );
